+++ src/main/java/org/swarg/mc/traprescue/Reference.java
@@ -5,5 +5,5 @@ package org.swarg.mc.traprescue;
 public class Reference {
     public static final String MOD_ID = "traprescue";
     public static final String MOD_NAME = "TrapRescue";
-    public static final String MOD_VERSION = "0.1.0";
+    public static final String MOD_VERSION = "0.2.0";
 }